Move Edit/Preview toggle into the workspace secondary header
The Edit/Preview toggle is a workspace-mode switch, not a branch/PR
action. Extract it into a standalone `ModeToggle` and render it on the
right side of the secondary header alongside the workspace cloud
features. `EditActions` now only carries Commit, Open PR, and Exit.
